Adding NDCG tests

+++ python/ml4ir/base/model/scoring/scoring_model.py
@@ -307,7 +307,7 @@ class RelevanceScorer(keras.Model):
             mask = inputs[self.feature_config.get_mask("node_name")]
         except:
             mask = None
-        self.compiled_metrics.update_state(y_true, y_pred, mask)
+        self.compiled_metrics.update_state(tf.cast(y_true, tf.float32), y_pred, mask)
 
         # Compute metrics on auxiliary label
         if self.aux_label:
